pyproject.toml
[project]
name = "ai-security-control-plane"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "openai>=2.9.0",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "python-dotenv==1.0.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "structlog>=25.5.0",
    "uvicorn>=0.46.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
